change dir base

+++ scripts/filelists/createFilelist.py
@@ -3,27 +3,30 @@
 import os
 from collections import OrderedDict
 
+base="/lustre/cmsdata/hoh/latino"
+#base="/media/shoh/02A1ACF427292FC0/nanov5"
+
 dirlist=OrderedDict({
     '2016' : {
-        'MCdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Summer16_102X_nAODv5_Full2016v6/MCl1loose2016v6__MCCorr2016v6__l2loose__l2tightOR2016v6" ,
-        'DATAdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Run2016_102X_nAODv5_Full2016v6/DATAl1loose2016v6__l2loose__l2tightOR2016v6" ,
-        'FAKEdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Run2016_102X_nAODv5_Full2016v6_ForNewWPs/DATAl1loose2016v6__l2loose__fakeW"
+        'MCdir' : base + "/Summer16_102X_nAODv5_Full2016v6/MCl1loose2016v6__MCCorr2016v6__l2loose__l2tightOR2016v6" ,
+        'DATAdir' : base + "/Run2016_102X_nAODv5_Full2016v6/DATAl1loose2016v6__l2loose__l2tightOR2016v6" ,
+        'FAKEdir' : base + "/Run2016_102X_nAODv5_Full2016v6_ForNewWPs/DATAl1loose2016v6__l2loose__fakeW"
     },
     '2017' : {
-        'MCdir'	: "/media/shoh/02A1ACF427292FC0/nanov5/Fall2017_102X_nAODv5_Full2017v6/MCl1loose2017v6__MCCorr2017v6__l2loose__l2tightOR2017v6" ,
-        'DATAdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Run2017_102X_nAODv5_Full2017v6/DATAl1loose2017v6__l2loose__l2tightOR2017v6" ,
-        'FAKEdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Run2017_102X_nAODv5_Full2017v6/DATAl1loose2017v6__l2loose__l2tightOR2017v6"
+        'MCdir'	: base + "/Fall2017_102X_nAODv5_Full2017v6/MCl1loose2017v6__MCCorr2017v6__l2loose__l2tightOR2017v6" ,
+        'DATAdir' : base + "/Run2017_102X_nAODv5_Full2017v6/DATAl1loose2017v6__l2loose__l2tightOR2017v6" ,
+        'FAKEdir' : base + "/Run2017_102X_nAODv5_Full2017v6/DATAl1loose2017v6__l2loose__l2tightOR2017v6"
     },
     '2018' : {
-        'MCdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Autumn18_102X_nAODv6_Full2018v6/MCl1loose2018v6__MCCorr2018v6__l2loose__l2tightOR2018v6" ,
-        'DATAdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Run2018_102X_nAODv6_Full2018v6/DATAl1loose2018v6__l2loose__l2tightOR2018v6" ,
-        'FAKEdir' : "/media/shoh/02A1ACF427292FC0/nanov5/Run2018_102X_nAODv6_Full2018v6_ForNewWPs/DATAl1loose2018v6__l2loose__fakeW"
+        'MCdir' : base + "/Autumn18_102X_nAODv6_Full2018v6/MCl1loose2018v6__MCCorr2018v6__l2loose__l2tightOR2018v6" ,
+        'DATAdir' : base + "/Run2018_102X_nAODv6_Full2018v6/DATAl1loose2018v6__l2loose__l2tightOR2018v6" ,
+        'FAKEdir' : base + "/Run2018_102X_nAODv6_Full2018v6_ForNewWPs/DATAl1loose2018v6__l2loose__fakeW"
     }
 })
 
 for ilist in dirlist:
     year=ilist ; MCdir=dirlist[ilist]['MCdir'] ; DATAdir=dirlist[ilist]['DATAdir'] ; FAKEdir=dirlist[ilist]['FAKEdir']
-
+    if year=="2016": continue
     if os.path.isdir('%s/%s' %( os.getcwd() , year ) ):
         os.system('rm -r %s/%s' %( os.getcwd() , year ))
